ConsoleApplication1.cpp

+++ ConsoleApplication1.cpp
@@ -43,7 +43,7 @@ public:
 class bolme {
 public:
 	int sayi1, sayi2;
-	double bolum = 0;
+	float bolum = 0;
 
 	void bolme_islemi() {
 		this->bolum = this->sayi1 / this->sayi2;
